Delay checkBuffers_ calls
This I don't really want to have to think about checkBuffers_ being called in
the middle of a function. Ensure that they also get new event loop iterations.
Also gives us a place to throttle it. (And setTimeout is throttled anyway, so
we got it for free.)

Particularly I'm uneasy about it being triggered on scroll and whether any of
those events are synchronous.