Rework mobile tap detection.
Now we only use hammer.js on iOS, where we can't use click events due to
the 300ms delay inherent in Safari (it waits for a double-tap, and this
can't be disabled). Everywhere else, we use click events. Fixes #88,
hopefully for real this time.

(There is some latency on Android tablets, but I'm not sure how much of
that is an event delay and how much is just lag. Either way, it
still *works*.)